requirements.txt
aws-cdk-lib==2.167.2
constructs>=10.0.0,<11.0.0
aws-solutions-constructs.aws-cloudfront-s3
